+++ contracts/adminBases/interface/IAsset_.sol
@@ -7,36 +7,48 @@ import "../../interface/IAsset.sol";
 
 interface IAsset_ is IAsset, IOwnable {
     function registerAsset(
-        address assetToken, 
-        address assetOracle, 
-        uint16 auctionDiscount, 
-        uint16 minCRatio, 
-        uint16 targetRatio, 
-        bool isInPreIPO, 
-        uint poolId, 
+        address assetToken,
+        address assetOracle,
+        uint16 auctionDiscount,
+        uint16 minCRatio,
+        uint16 targetRatio,
+        bool isInPreIPO,
+        uint256 poolId,
         IPOParams memory ipoParams
     ) external;
 
     function updateAsset(
-        address assetToken, 
-        address assetOracle, 
-        uint16 auctionDiscount, 
-        uint16 minCRatio, 
-        uint16 targetRatio, 
-        bool isInPreIPO, 
-        uint poolId, 
+        address assetToken,
+        address assetOracle,
+        uint16 auctionDiscount,
+        uint16 minCRatio,
+        uint16 targetRatio,
+        bool isInPreIPO,
+        uint256 poolId,
         IPOParams memory ipoParams
     ) external;
 
-    function registerCollateral(address cAssetToken, address oracle, uint16 multiplier) external;
+    function registerCollateral(
+        address cAssetToken,
+        address oracle,
+        uint16 multiplier
+    ) external;
 
-    function updateCollateral(address cAssetToken, address oracle, uint16 multiplier) external;
+    function updateCollateral(
+        address cAssetToken,
+        address oracle,
+        uint16 multiplier
+    ) external;
 
     function revokeCollateral(address cAssetToken) external;
 
     function triggerIPO(address assetToken) external;
 
-    function registerMigration(address assetToken, uint endPrice, uint8 endPriceDecimals) external;
+    function registerMigration(
+        address assetToken,
+        uint256 endPrice,
+        uint8 endPriceDecimals
+    ) external;
 
     function setCollateralInPreIPO(address cAssetToken, bool value) external;
 }
