+++ contracts/interface/IPositions.sol
@@ -4,15 +4,15 @@ pragma solidity ^0.8.2;
 
 import "./IAssetToken.sol";
 
-struct Position{
-    uint id;
+struct Position {
+    uint256 id;
     address owner;
     // collateral asset token.
     IERC20Extented cAssetToken;
-    uint cAssetAmount;
+    uint256 cAssetAmount;
     // nAsset token.
     IAssetToken assetToken;
-    uint assetAmount;
+    uint256 assetAmount;
     // if is it short position
     bool isShort;
     // 判断该空间是否已被分配
@@ -23,17 +23,26 @@ interface IPositions {
     function openPosition(
         address owner,
         IERC20Extented cAssetToken,
-        uint cAssetAmount,
+        uint256 cAssetAmount,
         IAssetToken assetToken,
-        uint assetAmount,
+        uint256 assetAmount,
         bool isShort
-    ) external returns(uint positionId);
+    ) external returns (uint256 positionId);
 
     function updatePosition(Position memory position_) external;
 
-    function removePosition(uint positionId) external;
+    function removePosition(uint256 positionId) external;
 
-    function getPosition(uint positionId) external view returns(Position memory);
-    function getNextPositionId() external view returns(uint);
-    function getPositions(address ownerAddr, uint startAt, uint limit) external view returns(Position[] memory);
+    function getPosition(uint256 positionId)
+        external
+        view
+        returns (Position memory);
+
+    function getNextPositionId() external view returns (uint256);
+
+    function getPositions(
+        address ownerAddr,
+        uint256 startAt,
+        uint256 limit
+    ) external view returns (Position[] memory);
 }
