+++ contracts/interface/IShortLock.sol
@@ -5,17 +5,28 @@ pragma solidity ^0.8.2;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 struct PositionLockInfo {
-    uint positionId;
+    uint256 positionId;
     address receiver;
     IERC20 lockedToken; // address(1) means native token, such as ETH or MITIC.
-    uint lockedAmount;
-    uint unlockTime;
+    uint256 lockedAmount;
+    uint256 unlockTime;
     bool assigned;
 }
 
 interface IShortLock {
-    function lock(uint positionId, address receiver, address token, uint amount) external payable;
-    function unlock(uint positionId) external;
-    function release(uint positionId) external;
-    function lockInfoMap(uint positionId) external view returns(PositionLockInfo memory);
+    function lock(
+        uint256 positionId,
+        address receiver,
+        address token,
+        uint256 amount
+    ) external payable;
+
+    function unlock(uint256 positionId) external;
+
+    function release(uint256 positionId) external;
+
+    function lockInfoMap(uint256 positionId)
+        external
+        view
+        returns (PositionLockInfo memory);
 }
